+++ test/pipeline_test.rb
@@ -258,4 +258,12 @@ class PipelineTest < Minitest::Test
     assert_elements_in_delta [0.067657, 0.063496, 0.048713], embeddings[0][..2]
     assert_elements_in_delta [0.086439, 0.10276, 0.0053946], embeddings[1][..2]
   end
+
+  def test_dtype
+    sentences = ["This is an example sentence", "Each sentence is converted"]
+    embed = Informers.pipeline("embedding", "Xenova/all-MiniLM-L6-v2", dtype: "fp16")
+    embeddings = embed.(sentences)
+    assert_elements_in_delta [0.067657, 0.063496, 0.048713], embeddings[0][..2]
+    assert_elements_in_delta [0.086439, 0.10276, 0.0053946], embeddings[1][..2]
+  end
 end